console: Move sumbit logic to container
