+++ pkg/webui/locales/en.json
@@ -685,6 +685,8 @@
   "console.containers.pubsub-formats-select.index.warning": "Pub/Sub formats unavailable",
   "console.containers.pubsubs-table.index.format": "Format",
   "console.containers.pubsubs-table.index.host": "Server host",
+  "console.containers.user-data-form.edit.updateSuccess": "User updated",
+  "console.containers.user-data-form.edit.deleteSuccess": "User deleted",
   "console.containers.users-table.index.invite": "Invite user",
   "console.containers.users-table.index.revokeInvitation": "Revoke this invitation",
   "console.containers.users-table.index.sentAt": "Sent",
@@ -822,8 +824,6 @@
   "console.views.admin-packet-broker.messages.defaultGatewayVisibilitySet": "Default gateway visibility set",
   "console.views.admin-packet-broker.messages.packetBrokerStatusPage": "Packet Broker Status page",
   "console.views.admin-panel-network-information.index.title": "Network information",
-  "console.views.admin-user-management-edit.index.updateSuccess": "User updated",
-  "console.views.admin-user-management-edit.index.deleteSuccess": "User deleted",
   "console.views.application-add.index.appDescription": "Within applications, you can register and manage end devices and their network data. After setting up your device fleet, use one of our many integration options to pass relevant data to your external services.{break}Learn more in our guide on <Link>Adding Applications</Link>.",
   "console.views.application-data.index.appData": "Application data",
   "console.views.application-integrations-lora-cloud.index.loraCloudInfoText": "Lora Cloud provides value added APIs that enable simple solutions for common tasks related to LoRaWAN networks and LoRa-based devices. You can setup our LoRaCloud integrations below.",
